Implement Flavors
A "flavor" is a custom set of MSBuild properties which can be used in the
Cake.BuildUtil Build Step. This could for example be used to specify different
upload locations for a ClickOnce application based on a flavor string
(e.g. while "Production" would push the ClickOnce application to server X,
"Beta" would push the application to server Y)