Use hardcoded bin folders again
The current ParseProject() Cake method cannot work with run configurations
that are not Debug. The "fixed" parse method is included in the Incubator
addin, so it might take some time before it is integrated into Cake. The
current solution is not ideal but is at least more reliant than using
ParseProjects().